Updated roach_spead_simulators to correctly transmit pointer/immediate bit. Added SWVER, INSTRUME fields, removed BACKEND, OBSERVER. Fixed packet loss counter.  Please enter the commit message for your changes. Lines starting

+++ src/guppi_daq/python/guppi_set_params.py
@@ -103,7 +103,7 @@ add_param_option("--npol",
         name="NPOL", type="int",
         help="Number of hardware polarizations")
 add_param_option("--nchan", 
-        name="NCHAN", type="float",
+        name="NCHAN", type="int",
         help="Number of spectral channels per sub-band")
 add_param_option("--chan_bw", 
         name="CHAN_BW", type="float",
